fix(BA-6913): stop creating the vfolders.creator_id index nobody reads
Converging the two schema paths only makes sense in the direction the index
earns. vfolders.creator_id carries no foreign key and is never filtered,
joined, or ordered by anywhere in the tree -- it is only written and
selected -- so the index it got in b4e7f1a2c3d5 costs a write on every
vfolder and saves nothing. Rather than declare it on the model to match,
stop creating it, and drop it from the databases that already ran that
revision.

The reconcile migration now moves each index in whichever direction it
earns: it creates the two that back a foreign key whose parent-side delete
must find the referencing rows, and drops this one.
