+++ src/ai/backend/manager/models/alembic/versions/b4e7f1a2c3d5_add_creator_id_to_vfolders.py
@@ -23,12 +23,6 @@ def upgrade() -> None:
         "vfolders",
         sa.Column("creator_id", GUID, nullable=True),
     )
-    op.create_index(
-        op.f("ix_vfolders_creator_id"),
-        "vfolders",
-        ["creator_id"],
-        unique=False,
-    )
     # Backfill creator_id from users table by matching creator (email) → users.email
     op.execute(
         sa.text(
